# Optimize isSpace functions