GHC unlike GCC does not optimize expressions like `x == 10 || x == 11 || x == 12 || x == 13` into `x >= 10 && x <= 13` and further into `x - 10 <= 3` so I did it here manually. Turns out this optimization was already applied years ago to `Data.Char.isSpace`: https://hackage.haskell.org/package/base-4.12.0.0/docs/src/GHC.Unicode.html#isSpace

I chose the `w == 0x20 || w == 0xA0 || w - 0x09 <= 4` order or terms instead of `w == 0x20 || w - 0x09 <= 4 || w == 0xA0` because it was faster on my machine. It also uses one fewer register according to the Compiler Explorer. It might be beneficial to adopt this order of terms in `Data.Char.isSpace` as well.

I also added a benchmark for `words` that uses `isSpaceWord8` a lot.

Before:

```
benchmarked words/lots of words
time                 142.2 μs   (141.8 μs .. 142.7 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 141.9 μs   (141.7 μs .. 142.2 μs)
std dev              746.9 ns   (611.1 ns .. 927.7 ns)

benchmarked words/one huge word
time                 11.73 μs   (11.71 μs .. 11.75 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 11.74 μs   (11.72 μs .. 11.76 μs)
std dev              62.67 ns   (46.41 ns .. 86.93 ns)

```

After:

```
benchmarked words/lots of words
time                 133.1 μs   (132.7 μs .. 133.7 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 133.1 μs   (132.8 μs .. 133.5 μs)
std dev              1.003 μs   (671.2 ns .. 1.578 μs)

benchmarked words/one huge word
time                 10.11 μs   (10.08 μs .. 10.13 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 10.10 μs   (10.08 μs .. 10.13 μs)
std dev              71.64 ns   (50.03 ns .. 116.5 ns)

```