@ethercrow 2020-10-25
For the curious, here is how GCC compiles `isspace`, no jumps at all: https://godbolt.org/z/zb1oso