@Bodigrim 2020-10-25
GCC implementation is equivalent to 
```haskell
import Data.Bits

isSpace :: Word8 -> Bool 
isSpace x = x - 0x09 <= 4 || x .&. 0x7f == 0x20
```

and one can use `GHC.Exts.{isTrue,leWord,eqWord,and,or}#` to implement it in a jumpless way.