@ethercrow 2020-10-25
Thanks for the hint, I just tried this:

```
isSpaceWord8 :: Word8 -> Bool
isSpaceWord8 (W8# w) =
  isTrue# (orI#
    (eqWord# (xor# w 0x7f##) 0xa0##)       -- ' ' or nbsp
    (leWord# (minusWord# w 0x09##) 4##))  -- \t, \n, \v, \f, \r
```

It slowed down `words/lots of words` from 130-ish to 150-ish us, looks like early exit on the space character is really valuable. There was no noticable effect on `words/one huge word`.

If I add the 0x20 check to this version, it performs identically to what is already in PR. So it's probably better not to add all those magic hashes.