@Bodigrim 2020-10-25
Well, your benchmark is obviously biased in favor of the short-circuiting version, but I guess it is not different in this aspect from the real-world data. Recently @vdukhovni and I discussed a similar function in https://github.com/haskell-streaming/streaming-bytestring/pull/31#discussion_r497713778