@vdukhovni 2020-10-25
> Well, your benchmark is obviously biased in favor of the short-circuiting version, but I guess it is not different in this aspect from the real-world data. Recently @vdukhovni and I discussed a similar function in [haskell-streaming/streaming-bytestring#31 (comment)](https://github.com/haskell-streaming/streaming-bytestring/pull/31#discussion_r497713778)

Indeed the proposed function is almost identical to the one in streaming bytestring, except that I optimise for most characters being non-whitespace ASCII characters, by first ruling out most of those:
```haskell
-- Predicate to test whether a 'Word8' value is either ASCII whitespace,
-- or a unicode NBSP (U+00A0).  Optimised for ASCII text, with spaces
-- as the most frequent whitespace characters.
w8IsSpace :: Word8 -> Bool
w8IsSpace = \ !w8 ->
    -- Avoid the cost of narrowing arithmetic results to Word8,
    -- the conversion from Word8 to Word is free.
    let w :: Word
        !w = fromIntegral w8
     in w - 0x21 > 0x7e   -- not [x21..0x9f]
        && ( w == 0x20    -- SP
          || w - 0x09 < 5 -- HT, NL, VT, FF, CR
          || w == 0xa0 )  -- NBSP
{-# INLINE w8IsSpace #-}
```
I am curious how the above compares with this PR on "real world" test data...