@ethercrow 2020-10-26
Changed the test case from `intersperse ' ' bigData` to lorem ipsum.

BTW how do you run only a subset of benchmarks with `cabal bench`? I'm just deleting everything not relevant for a moment, that's rather cumbersome.