@sjakobi 2020-10-26
> BTW how do you run only a subset of benchmarks with `cabal bench`?

You can pass in options via `--benchmark-option[s]`, e.g.

```
$ cabal bench bench-bytestring-builder --benchmark-option folds/scanl/1
```

Try passing in `-h` to see the various CLI options – `--match` might be useful.

(https://github.com/vincenthz/hs-gauge/issues/97 is related)