@ethercrow 2020-10-26
Viktor's version is the winner on lorem ipsum on my machine as well. So let's adopt that.