@Bodigrim 2020-10-28
@ethercrow Tests for GHC < 7.10 are failing.
https://travis-ci.org/github/haskell/bytestring/jobs/739381052#L285
```
Data/ByteString/Internal.hs:679:14:
    Not in scope: type constructor or class ‘Word’
    Perhaps you meant ‘Word8’ (imported from Data.Word)
```