@vdukhovni 2020-10-28
For compatibility with older GHC, you'll need to import `Word` from `Data.Word` into `Data.ByteString.Internal`:
old:
```haskell
import Data.Word                (Word8)
```
new:
```haskell
import Data.Word                (Word8, Word)
```
With that, the CI tests should pass.