@ethercrow 2020-10-29
> @ethercrow while we are waiting for @sjakobi to review, do you want me to label this as "hacktoberfest-accepted"?

I started looking at bytestring when you posted a hacktoberfest call to arms, but that hacktoberfest context was not important to me, so don't worry about it. Thank you for caring though!